+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- docspace: added the DocSpace plugins SDK
 - docspace: storybook v2.0.0
 - editors: added the information about Health check
 - desktop: added the portal:uiTheme command to the execCommand method